Frontend: VideoPlayer: Workaround for Chromium desync
When playing a video in the background, then going back to the tab of the video, audio and video will de-sync.

As a workaround, when document.visibilityState changes to 'visible', the player will set videoElement.currentTime to force the media engine to re-sync the audio and the video.

Only happens in Chromium browsers.

This workaround should be removed as soon as Chromium fixes this issue.